# app.py
from flask import Flask, render_template, redirect, url_for, request, jsonify
from flask_sqlalchemy import SQLAlchemy
from datetime import datetime
import os
from dotenv import load_dotenv
from flask_swagger_ui import get_swaggerui_blueprint
from werkzeug.security import generate_password_hash, check_password_hash
from flask_jwt_extended import create_access_token, get_jwt_identity, jwt_required, JWTManager

app = Flask(__name__)
#***********************|DATABASE CONNECTION START|*********************#
load_dotenv()
app.config['SQLALCHEMY_DATABASE_URI'] = os.environ.get('DATABASE_URL')
app.config['SQLALCHEMY_TRACK_MODIFICATIONS'] = False
db = SQLAlchemy(app)
#######################|DATABASE CONNECTION END|##########################


#***********************|JWT MANAGER START|*********************#
app.config['JWT_SECRET_KEY'] = os.environ.get('JWT_SECRET_KEY')
jwt = JWTManager(app)
#######################|JWT MANAGER END|##########################


#***********************|SWAGGER DOC START|*********************#
SWAGGER_URL = '/api/docs'
API_URL = '/static/swaggerDoc1.3.json'  
swaggerui_blueprint = get_swaggerui_blueprint(SWAGGER_URL, API_URL, config={'app_name': "Test application"},)
app.register_blueprint(swaggerui_blueprint)

#######################|SWAGGER DOC END|##########################

@app.route('/')
def Home():
    return "Home page"
#***********************|LOGIN START|*********************#

@app.route('/login', methods=['POST'])
def login():
    '''
    User Login using JWT authentication
    '''
    if not request.is_json:
        return {"error": "Missing JSON in request"}, 400
    data = request.get_json()
    username = data.get('username')
    password = data.get('password')
    user = User.query.filter_by(username=username).first()
    if not user or not user.check_password(password):
        return {"error": "Wrong Password or Username"}, 401

    access_token = create_access_token(identity=user.username)
    return {"access_token": access_token}, 200

@app.route('/protected', methods=['GET'])
@jwt_required()
def protected():
    '''
    A protected endpoint that requires a valid JWT to access
    '''
    current_user = get_jwt_identity() 
    if not current_user:
        return {"error": "User not found"}, 404

    return jsonify(logged_in_as=current_user), 200

#######################|LOGIN END|##########################

#***********************|USER ENDPOINTS START|*********************#
@app.route('/createUser', methods=['POST'])
def createUser():
    '''
    Create a new user and save in DB
    '''
    if not request.is_json:
        return {"error": "Invalid input, send user details in JSON format"}, 400
    
    data = request.get_json()
    try:
        user = User(
            username=data.get('username'),
            email=data.get('email')
        )
        user.set_password(data.get('password'))
        db.session.add(user)
        db.session.commit()
        return {"message": "User created", "user_id": user.id}, 201
    except Exception as e:
        db.session.rollback()
        return {"error": str(e)}, 400
        

@app.route('/getUsers', methods=['GET'])
def getUsers():
    '''
    Get all users from DB
    '''
    try:
        users = User.query.all()
        users_list = [{"id": user.id, "username": user.username, "email": user.email} for user in users]
        return {"users": users_list}, 200
    except Exception as e:
        return {"error": str(e)}, 400


@app.route('/getUser/<id>', methods=['GET'])
def getUserByID(id):
    '''
    Get existing user from DB using user ID
    '''
    try:
        user = User.query.get(id)
        if user:
            return {"id": user.id, "username": user.username, "email": user.email}, 200
        else:
            return {"error": "User not found"}, 404
    except Exception as e:
        return {"error": str(e)}, 400


#################|USER ENDPOINTS END|#####################

#***********************|Client ENDPOINTS START|*********************#



#################|Client ENDPOINTS END|#####################

@app.route('/createClient', methods=['POST'])
def createClient():
    '''
    Create a new Client and save in DB
    '''
    if not request.is_json:
        return {"error": "Invalid input, send user details in JSON format"}, 400
    
    data = request.get_json()
    try:
        client = Client(
            username=data.get('username'),
            email=data.get('email')
        )
        db.session.add(client)
        db.session.commit()
        return {"message": "Client created", "client_id": client.id}, 201
    except Exception as e:
        db.session.rollback()
        return {"error": str(e)}, 400
        

@app.route('/getClients', methods=['GET'])
def getClients():
    '''
    Get all Clients from DB
    '''
    try:
        clients = Client.query.all()
        clients_list = [{"id": client.id, "username": client.username, "email": client.email} for client in clients]
        return {"client": clients_list}, 200
    except Exception as e:
        return {"error": str(e)}, 400


@app.route('/getClient/<id>', methods=['GET'])
def getClientByID(id):
    '''
    Get existing client from DB using user ID
    '''
    try:
        client = Client.query.get(id)
        if client:
            return {"id": client.id, "username": client.username, "email": client.email}, 200
        else:
            return {"error": "Client not found"}, 404
    except Exception as e:
        return {"error": str(e)}, 400



#***********************|CONTRACT ENDPOINTS START|*********************#

@app.route('/createContract', methods=['POST'])
def createContract():
    '''
    Create a new contract and save in DB
    '''
    if not request.is_json:
        return {"error": "Invalid input, send contract details in JSON format"}, 400
    
    # Extract data from json request
    data = request.get_json()

    contract = Contract(
        #contract_id=data.get('contract_id'),
        client_id=data.get('client_id'),
        api_id=data.get('api_id'),
        contract_type=data.get('contract_type'),
        pricing_type=data.get('pricing_type'),
        start_date=datetime.strptime(data.get('start_date'), '%Y-%m-%d'),
        end_date=datetime.strptime(data.get('end_date'), '%Y-%m-%d'),
        contract_status=data.get('contract_status', 'Draft'),
        value=float(data.get('value', 0.0))
    )
    db.session.add(contract)
    db.session.commit()
    return {"message": "Contract created", "contract_id": contract.contract_id}, 201 

@app.route('/archiveContract', methods=['DELETE'])
def deleteContractByID():
    '''
    Delete existing contract from DB using contract ID
    '''
    pass 

@app.route('/updateContract/{contractID}', methods=['PUT'])
def updateContractByID(contractID):
    '''
    Update existing contract in DB using contract ID
    '''
    try:
        contract = Contract.query.get(contractID)
        if not contract:
            return {"error": "Contract not found"}, 404
        if not request.is_json:
            return {"error": "Invalid input, send contract details in JSON format"}, 400
        data = request.get_json()
        contract.client_id = data.get('client_id', contract.client_id)
        contract.api_id = data.get('api_id', contract.api_id)
        contract.contract_type = data.get('contract_type', contract.contract_type)
        contract.pricing_type = data.get('pricing_type', contract.pricing_type)
        contract.start_date = datetime.strptime(data.get('start_date'), '%Y-%m-%d') if data.get('start_date') else contract.start_date
        contract.end_date = datetime.strptime(data.get('end_date'), '%Y-%m-%d') if data.get('end_date') else contract.end_date
        contract.contract_status = data.get('contract_status', contract.contract_status)
        contract.value = float(data.get('value', contract.value))
        db.session.commit()
        return {"message": "Contract updated", "contract_id": contract.contract_id}, 200
    except Exception as e:
        db.session.rollback()
        return {"error": str(e)}, 400


    

@app.route('/getContracts', methods=['GET'])
def getAllContracts():
    '''
    Get all existing contracts from DB
    '''
    try:
        contracts = Contract.query.all()
        contracts_list = [{
            "contract_id": contract.contract_id,
            "client_id": contract.client_id,
            "api_id": contract.api_id,
            "contract_type": contract.contract_type,
            "pricing_type": contract.pricing_type,
            "start_date": contract.start_date.strftime('%Y-%m-%d'),
            "end_date": contract.end_date.strftime('%Y-%m-%d'),
            "contract_status": contract.contract_status,
            "value": contract.value
        } for contract in contracts]
        return {"contracts": contracts_list}, 200

    except Exception as e:
        return {"error": str(e)}, 400

@app.route('/getContract/<contractID>', methods=['GET'])
def getContractByID(contractID):
    '''
    Get existing contract from DB using contract ID
    '''
    try:
        contract = Contract.query.get(contractID)
        if contract:
            return {
                "contract_id": contract.contract_id,
                "client_id": contract.client_id,
                "api_id": contract.api_id,
                "contract_type": contract.contract_type,
                "pricing_type": contract.pricing_type,
                "start_date": contract.start_date.strftime('%Y-%m-%d'),
                "end_date": contract.end_date.strftime('%Y-%m-%d'),
                "contract_status": contract.contract_status,
                "value": contract.value
            }, 200
        else:
            return {"error": "Contract not found"}, 404
    except Exception as e:
        return {"error": str(e)}, 400
     

@app.route('/getContractsByUser/<userID>', methods=['GET'])
def getContractByUserID(userID):
    '''
    Get existing contract from DB using Client ID
    '''
    try:
        contracts = Contract.query.filter_by(client_id=userID).all()
        contracts_list = [{
            "contract_id": contract.contract_id,
            "client_id": contract.client_id,
            "api_id": contract.api_id,
            "contract_type": contract.contract_type,
            "pricing_type": contract.pricing_type,
            "start_date": contract.start_date.strftime('%Y-%m-%d'),
            "end_date": contract.end_date.strftime('%Y-%m-%d'),
            "contract_status": contract.contract_status,
            "value": contract.value
        } for contract in contracts]
        return {"contracts": contracts_list}, 200

    except Exception as e:
        return {"error": str(e)}, 400

@app.route('/getContractsByProduct/<productID>', methods=['GET'])
def getContractsByProductID(productID):
    '''
    Get existing contract from DB using Client ID
    '''
    try:
        contracts = Contract.query.filter_by(api_id=productID).all()
        contracts_list = [{
            "contract_id": contract.contract_id,
            "client_id": contract.client_id,
            "api_id": contract.api_id,
            "contract_type": contract.contract_type,
            "pricing_type": contract.pricing_type,
            "start_date": contract.start_date.strftime('%Y-%m-%d'),
            "end_date": contract.end_date.strftime('%Y-%m-%d'),
            "contract_status": contract.contract_status,
            "value": contract.value
        } for contract in contracts]
        return {"contracts": contracts_list}, 200

    except Exception as e:
        return {"error": str(e)}, 400


#################|CONTRACT ENDPOINTS END|#####################

#***********************|PRODUCTS(APIs) START|*********************#

@app.route('/createProduct', methods=['POST'])
def createProduct():
    '''
    create a new API product and add to DB
    '''
    if not request.is_json:
        return {"error": "Invalid input, send product details in JSON format"}, 400
    data = request.get_json()
    try:
        product = Product(
            name=data.get('name'),
            version=data.get('version'),
            pricing_type=data.get('pricing_type'),
            price_per_month=float(data.get('price', 0.0)),
            calls_per_month=int(data.get('calls_per_month', 0)),
            call_limit_type=int(data.get('call_limit_type'))
        )
        db.session.add(product)
        db.session.commit()
        return {"message": "Product created", "product_id": product.id}, 201

    except Exception as e:
        db.session.rollback()
        return {"error": str(e)}, 400


@app.route('/getProducts', methods=['GET'])
def getProducts():
    '''
    Get all API products from DB
    '''
    try:
        products = Product.query.all()
        products_list = [{
            "id": product.id,
            "name": product.name,
            "version": product.version,
            "pricing_type": product.pricing_type,
            "price_per_month": product.price_per_month,
            "calls_per_month": product.calls_per_month,
            "call_limit_type": product.call_limit_type
        } for product in products]
        return {"products": products_list}, 200

    except Exception as e:
        return {"error": str(e)}, 400

@app.route('/getProduct',methods=['GET'])
def getProductByID():
    '''
    Get existing API product from DB using product ID
    '''
    pass

@app.route('/updateProduct', methods=['PUT'])
def updateProductByID():
    '''
    Update existing API product in DB using product ID
    '''
    pass

@app.route('/deleteProduct', methods=['DELETE'])
def deleteProductByID():
    '''
    Delete existing API product from DB using product ID
    '''
    pass

#########################|PRODUCTS(APIs) END|#####################



#***********************|DATABASE MODELS START|*********************#
class Client(db.Model):
    """
    Client.who sign contracts for APIs with the company
    """
    id = db.Column(db.Integer, primary_key=True, unique=True)
    username = db.Column(db.String(20), unique=True, nullable=False)
    email = db.Column(db.String(60), unique=True, nullable=False)

    contracts = db.relationship('Contract', backref='client', lazy=True)

    def __repr__(self):
        return f'<User {self.username}>'


class User(db.Model):
    """
    Employees who input contracts into the system
    """
    id = db.Column(db.Integer, primary_key=True, unique=True)
    username = db.Column(db.String(80), unique=True, nullable=False)
    email = db.Column(db.String(120), unique=True, nullable=False)
    password_hash = db.Column(db.String(256), nullable=False)

    def set_password(self, password):
        self.password_hash = generate_password_hash(password)

    def check_password(self, password):
        return check_password_hash(self.password_hash, password)

    def __repr__(self):
        return f'<User {self.username}>'

# ---

class Contract(db.Model):
    """
    Contract between clients and the company.
    """
    contract_id = db.Column(db.Integer, primary_key=True, unique=True) 
    client_id = db.Column(db.Integer, db.ForeignKey('client.id'), nullable=False)
    api_id = db.Column(db.Integer,db.ForeignKey('product.id') ,nullable=False)
    contract_type = db.Column(db.String(50))
    pricing_type = db.Column(db.String(50))
    start_date = db.Column(db.Date, nullable=False)
    end_date = db.Column(db.Date, nullable=False)
    contract_status = db.Column(db.String(20), default='Draft', nullable=False)
    value = db.Column(db.Float, default=0.0)
    created_at = db.Column(db.DateTime, default=datetime.utcnow)
    updated_at = db.Column(db.DateTime, default=datetime.utcnow, onupdate=datetime.utcnow)



    def __repr__(self):
        return f'<Contract {self.contract_id}>'

class Product(db.Model):
    """
    API products from company
    """
    id = db.Column(db.Integer, primary_key=True, unique=True)
    name = db.Column(db.String(50), unique=True, nullable=False)
    version = db.Column(db.String(5), nullable=False)
    pricing_type = db.Column(db.String(50), nullable=False)
    price_per_month = db.Column(db.Float, default=0.0)
    calls_per_month = db.Column(db.Integer, default=0)
    call_limit_type = db.Column(db.Integer, nullable=False)


    def __repr__(self):
        return f'<Product {self.api_id} ({self.name},{self.version})>'

##################|DATABASE MODELS END|#####################


if __name__ == '__main__':
    with app.app_context():
        db.create_all()
    app.run(debug=True)