# static/swaggerDoc1.3.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API Management System (v1.2)",
    "description": "A management system for contracts made between company and clients.",
    "version": "1.3.0"
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter the JWT token obtained from the /login endpoint (used by employees/internal users)."
      }
    },
    "schemas": {
      "InternalUser": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "username": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email"
          }
        },
        "description": "Employee or internal system user."
      },
      "Client": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "username": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email"
          }
        },
        "description": "External client who signs contracts."
      },
      "Contract": {
        "type": "object",
        "properties": {
          "contract_id": {
            "type": "integer"
          },
          "client_id": {
            "type": "integer",
            "description": "Foreign key to Client ID."
          },
          "api_id": {
            "type": "integer",
            "description": "Foreign key to Product ID."
          },
          "contract_type": {
            "type": "string"
          },
          "pricing_type": {
            "type": "string"
          },
          "start_date": {
            "type": "string",
            "format": "date",
            "example": "2025-01-01"
          },
          "end_date": {
            "type": "string",
            "format": "date",
            "example": "2025-12-31"
          },
          "contract_status": {
            "type": "string",
            "default": "Draft"
          },
          "value": {
            "type": "number",
            "format": "float"
          }
        },
        "required": [
          "client_id",
          "api_id",
          "start_date",
          "end_date"
        ]
      },
...
    }
  },
  "paths": {
...
    "/login": {
      "post": {
        "summary": "Internal User Login (Employee)",
        "description": "Authenticates an internal user and returns a JWT access token.",
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string",
                    "format": "password"
                  }
                },
                "required": [
                  "username",
                  "password"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful login, returns JWT."
          },
          "401": {
            "description": "Wrong Password or Username."
          }
        }
      }
    },
    "/protected": {
      "get": {
        "summary": "Protected Test Endpoint",
        "description": "Requires a valid JWT for access (Internal User only).",
        "tags": [
          "Auth"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Access granted.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "logged_in_as": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized (Missing or invalid token)."
          }
        }
      }
    },
    "/createUser": {
      "post": {
        "summary": "Create Internal User (Employee)",
        "description": "Creates a new internal system user (employee).",
        "tags": [
          "Internal Users"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string",
                    "format": "email"
                  },
                  "password": {
                    "type": "string",
                    "format": "password"
                  }
                },
                "required": [
                  "username",
                  "email",
                  "password"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User successfully created."
          },
          "400": {
            "description": "Invalid input or database error."
          }
        }
      }
    },
...
    "/getUser/{id}": {
      "get": {
        "summary": "Get Internal User by ID",
        "description": "Retrieves a single internal user by their database ID.",
        "tags": [
          "Internal Users"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User details.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InternalUser"
                }
              }
            }
          },
          "404": {
            "description": "User not found."
          }
        }
      }
    },
    "/createClient": {
      "post": {
        "summary": "Create Client",
        "description": "Creates a new external client.",
        "tags": [
          "Clients"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string",
                    "format": "email"
                  }
                },
                "required": [
                  "username",
                  "email"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Client successfully created."
          },
          "400": {
            "description": "Invalid input or database error."
          }
        }
      }
    },
...
    "/getClient/{id}": {
      "get": {
        "summary": "Get Client by ID",
        "description": "Retrieves a single client by their database ID.",
        "tags": [
          "Clients"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Client details.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Client"
                }
              }
            }
          },
          "404": {
            "description": "Client not found."
          }
        }
      }
    },
    "/createContract": {
      "post": {
        "summary": "Create Contract",
        "description": "Creates a new contract.",
        "tags": [
          "Contracts"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "client_id": {
                    "type": "integer"
                  },
                  "api_id": {
                    "type": "integer",
                    "description": "ID of the API product."
                  },
                  "contract_type": {
                    "type": "string",
                    "example": "Annual"
                  },
                  "pricing_type": {
                    "type": "string",
                    "example": "Subscription"
                  },
                  "start_date": {
                    "type": "string",
                    "format": "date",
                    "example": "2025-01-01"
                  },
                  "end_date": {
                    "type": "string",
                    "format": "date",
                    "example": "2025-12-31"
                  },
                  "contract_status": {
                    "type": "string",
                    "default": "Draft"
                  },
                  "value": {
                    "type": "number",
                    "format": "float"
                  }
                },
                "required": [
                  "client_id",
                  "api_id",
                  "start_date",
                  "end_date"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Contract created successfully. Returns the auto-generated contract_id."
          },
          "400": {
            "description": "Invalid input or database error."
          }
        }
      }
    },
...
    "/updateContract/{contractID}": {
      "put": {
        "summary": "Update Contract",
        "description": "Updates an existing contract using its ID. Any field in the request body is optional.",
        "tags": [
          "Contracts"
        ],
        "parameters": [
          {
            "name": "contractID",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/Contract"
                  }
                ],
                "properties": {
                  "contract_id": {
                    "readOnly": true
                  }
                },
                "required": []
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Contract updated successfully."
          },
          "404": {
            "description": "Contract not found."
          },
          "400": {
            "description": "Invalid input or database error."
          }
        }
      }
    },
...
    "/getContract/{contractID}": {
      "get": {
        "summary": "Get Contract by ID",
        "description": "Retrieves a single contract by its ID.",
        "tags": [
          "Contracts"
        ],
        "parameters": [
          {
            "name": "contractID",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Contract details.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Contract"
                }
              }
            }
          },
          "404": {
            "description": "Contract not found."
          }
        }
      }
    },
    "/getContractsByUser/{userID}": {
      "get": {
        "summary": "Get Contracts by Client ID",
        "description": "Retrieves all contracts associated with a specific Client ID.",
        "tags": [
          "Contracts"
        ],
        "parameters": [
          {
            "name": "userID",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A list of contracts for the user."
          }
        }
      }
    },
    "/getContractsByProduct/{productID}": {
      "get": {
        "summary": "Get Contracts by Product ID",
        "description": "Retrieves all contracts associated with a specific Product ID.",
        "tags": [
          "Contracts"
        ],
        "parameters": [
          {
            "name": "productID",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A list of contracts for the product."
          }
        }
      }
    },
    "/createProduct": {
      "post": {
        "summary": "Create API Product",
        "description": "Creates a new API product/service offering.",
        "tags": [
          "Products"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "version": {
                    "type": "string",
                    "example": "v1.0"
                  },
                  "pricing_type": {
                    "type": "string",
                    "example": "Tiered"
                  },
                  "price_per_month": {
                    "type": "number",
                    "format": "float"
                  },
                  "calls_per_month": {
                    "type": "integer"
                  },
                  "call_limit_type": {
                    "type": "integer",
                    "description": "Type of call limit (e.g., 1 for Hard, 2 for Soft)."
                  }
                },
                "required": [
                  "name",
                  "version",
                  "pricing_type",
                  "call_limit_type"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Product created successfully."
          },
          "400": {
            "description": "Invalid input or database error."
          }
        }
      }
    },
...
  }
}